+++ NEWS.md
@@ -6,10 +6,12 @@ Versioning](https://semver.org).
 
 ## Unreleased
 
-### Changed
+### Breaking changes
+
+- Out of range indices in lists of Tcl objects are no longer allowed.
 
 - `w.rgb(color)` with `w` a widget now return an `RGB{N0f16}` value instead of a triplet of
-   `UInt16`.
+  `UInt16`.
 
 ### Added